POKEGAME-38: implemented Move management. (#48)

+++ backend/src/PokeGame.Infrastructure/Converters/PowerPointsConverter.cs
@@ -6,7 +6,7 @@ internal class PowerPointsConverter : JsonConverter<PowerPoints>
 {
   public override PowerPoints? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
   {
-    return reader.TryGetInt32(out int value) ? new PowerPoints(value) : null;
+    return reader.TryGetByte(out byte value) ? new PowerPoints(value) : null;
   }
 
   public override void Write(Utf8JsonWriter writer, PowerPoints powerPoints, JsonSerializerOptions options)
